requirements.txt
requests>=2.32.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
feedparser>=6.0.0
python-dateutil>=2.9.0
schedule>=1.2.0
anthropic>=0.77.0
pyyaml>=6.0.0
